requirements.txt
SQLAlchemy~=2.0.32
fastapi~=0.112.2
jose~=1.0.0
passlib~=1.7.4
pydantic~=2.8.2
alembic~=1.13.2
fastapi-cache2[redis]